backend/requirements.txt
fastapi==0.103.2
uvicorn==0.25.2
google-genai==1.61.0
pydantic==2.12.5
anyio>=4.8.0,<4.10.0
typing-extensions>=4.11.0,<5.0.0
httpx>=0.28.1,<1.0.0
